Replaced ConfigOverride with EngineSupplement.
Motivation:
For an efficient component replication setup, we need to provide
the component types to the engine build. We were already doing
similar things through the ConfigOverride mechanism, but it
was getting stretched past its original intent, and didn't
have a way to provide cpp files.

This commit replaces ConfigOverride with a more complete mechanism:
EngineSupplement. With this, we're now adding files directly to the
engine's CMake targets in an organized way.

Changes:
Moved all override-based files to EngineSupplementExamples, to make
way for the project providing them.