Implemented item saving to database.
Changes:
Replaced ItemDataBase's item history vector with a signal, so
we can have multiple consumers.

Switched to an in-memory database for speed of initial copy from
our sim state. We then tell a thread to backup the in-memory db
to a file db.

SaveSystem now saves items.

World now loads items on startup.