Added maxStackSize to Items.
Changes:
Added support for setting scaling quality based on Config value.
Not sure if it was removed on purpose or never was present, but
it seemed like a good thing to do.

Added maxStackSize to Item, InventoryAddItem, CombineItems, and
ItemUpdate. The messages need it since the client may not have the
item definition when the message is processed.

Removed the "null item" to make item handling more intuitive.
NULL_ITEM_ID represents an empty space, and unlike sprites or
animations, we never need a "safe default" when it comes to
items. Furthermore, having an actual null item in the map was
complicating all of our item existence checks.

addDescription -> setDescription (+ it now overwrites).

Inventory now flows items when adding, based on maxStackSize.