Further dialogue fixes.
Changes:
Changed player entity names to "Player " + their entity ID
so it looks better in dialogues for now.

Fixed a lot of dialogue logic and bindings.

Added a range check to dialogue choices.

Added a narrate() binding, for narration text.

Added a choiceIf() binding, for the conditional version of choice().

Moved DialogueEvent into a separate header for easier use and sharing.