(Safety) Mouse picking, transform fixes.
Changes:
Started fixing mouse picking. Previously, we were using an abstract
step size that didn't account for the cell size : world unit ratio,
but it was fine since our cells were square. That broke with the
introduction of Z, but it seems fixed now.

Changed camera.zLength to be half a radius above the target, +
the distance from Z == 0 to the target. This gives us the uneven
Z-axis view distance that we actually want (see all the way to the
ground, + some distance above).

Started trying a different approach for screenToWorld, where we use the
closest tile Z under the camera's target as a base plane, and run calcs
relative to that.