Switched TileMapBase from grid to hash map.
Changes:
TileMapBase now uses a hash map to store chunks instead of a grid.
This introduced the possibility of a tile not existing, so null checks
were added to any usage of the map getters.

Updated some Tile functions to early-exit if they pass the layer type
they're looking for. This may not be any more efficient, or it may even
be less efficient, but it makes me feel better.

Moved asserts in TileMapBase into the generic functions to reduce code
footprint a bit. If it gets confusing not having them at the top, we can
move them back.

TileMapBase now manages the chunk's nonEmptyTileCount. If it becomes 0,
the chunk is erased.