Added offsets for Floor and Object tile layers.
Motivation:
Users are likely to desire placing Floors and Objects in non-
tile-aligned positions. To support this, we now have offsets that
let you move them by up to a tile's worth of distance.

Changes:
Reworked WorldSpriteSorter and PhantomSpriteInfo to pass the
necessary data through and apply it to both phantoms and normal
graphics.

Reworked map loading. Simplified some of it, made some of it more
efficient, and added support for offsets.

Reworked chunk streaming to support offsets, and to require fewer
loads while processing a chunk (switched from "vector per tile" to
parallel arrays).

Reworked ClientTransforms::tileToScreenExtent to support offsets,
and took the opportunity to also make it less one-off.

Reworked a lot of TileMapBase to support offsets. Also simplified the
Terrain interface.

Added support for getting an x,y coordinate, given a morton code. This
is useful for chunk loading, since we need to know the coordinate when
calculating bounds.