Started adding bounding box "save as".
Changes:
AnimationPropertiesWindow's bounding box button now
provides "Save as" functionality. This is part of a move away
from editing bounding boxes in their own view, to doing it in
sprite/animation context.
One reason for this change is that we have the stage graphic
now, but the bounding box view doesn't have an image size to
use when sizing the extent. Another reason is that the
sprite/animation-first workflow seems to be more useful.

Removed the animation view's sprite assignment button and
lowered things to make room for the stage graphic. Since
animations are going to be defined through regions soon,
the button won't be needed.

Removed the animation view's play button, added "space bar
to play", and made the view keyboard focusable.

Added SaveBoundingBoxDialog. Went a little lazy with the
layout and just re-used the sizing from another dialog.