(Safety) Fixes for adding sprites to animations.
Changes:
Sprites are now added to the proper slots.

Fixed some crashes.

We now check sprite size to make sure it matches the other
sprites in the animation before adding it.

Added an altenative representation for the animation frame
vector. It's more useful for certain operations, but the
existing form is still useful for when you only care about
the frames that actually exist.

If you select an empty animation slot, we now display the
sprite from a previous filled slot if there is one.

Fixed an issue where, when there was only 1 sprite in a sheet,
we were cutting it in half. That should be the only case where
we need to cut vertically, all others should be horizontal.

When you delete the last sprite from an animation, the animation
is now deleted.

You can now only reduce an animation's frame count up to the last
filled slot. This is to avoid more complicated removal logic, if you
want to make the animation smaller you'll have to move or remove
sprites first.