(Safety) Started implementing alignment anchors.
Changes:
Instead of assuming a stage size of 1 tile, entities now have their
bounding box centered on their position. Graphic logic has been
updated to match. Soon, this should be replaced with centering on
the graphic's alignment anchor.

We now use different calculations for model -> world for entity
bounding boxes, depending on whether they're for collision or for
render sorting. This lets us keep collision box updating simple,
while doing whatever we need to on the render side to keep the
graphic aligned.

Moved some entity graphic functionality into EntityGraphicSet. This
is more convenient to use, and enforces that IdleSouth is the graphic
to use for all sim bounding boxes.

Added alignment offset support to the engine. Still need to add to the
editor.

Added World::teleportEntity(), so project systems don't have to do it
manually and possibly forget a step.