(Safety) More alignment anchor changes.
Changes:
Switched from the alignmentOffset concept to alignment anchors, due
to realizing that we need to align a given entity graphic's anchor
with IdleSouth's bottom center. This makes more sense to do with an
anchor value, it'd be confusing to pre-compute it.

Decided to use the collision box (IdleSouth's modelBounds) for
sorting, instead of the current graphic's bounds. This is because
using a different box would open us up to having multiple boxes
clipping inside of eachother during sorting. It seems like a better
idea to just use the box that we know won't be clipping anything else,
and let the graphics sort however they end up.