Alignment anchors are working.
Changes:
Added editor support for alignment anchors.

Stage origin now updates as you go through graphics in an
animation in the editor. The alternative would have been to
make the engine use the stage origin from the first graphic in
the animation, which may be fine because there likely is
never a reason to have different stage origins between
graphics in the same animation. We decided to go this way
though, because it's more intuitive and "you only need to have
the correct stage origin on the first sprite" isn't a
feature that we care to support.