Fixed some editor bugs.
Changes:
Fixed a crash when deleting sprite sheets.

Fixed a bug where sprite sheets weren't selectable in the
library.

Fixed shift+selecting in the library.

Fixed ctrl+selecting to deselect items in the library (we
previously weren't accounting for this as a possible method
of deselection).

Fixed a bug where sets weren't getting updated when animations
were deleted.