Added premultiply alpha option for sprites.
Changes:
Added per-sprite option for premultiplying a sprite's alpha.
If checked, we do the premultiply while exporting sprite
sheets (the given texture should not be premultiplied).
We don't premultiply when showing the graphics in the importer,
that's a future improvement that could be made. The engine
looks at the option and handles them with the correct blend
mode, though.

We've standardized on ARGB8888 for texture generation. Any
given sprite textures that aren't already in that format will be
converted.

Added 2px of padding to sprites in sprite sheets, to get rid of
filtering artifacts.