Bugfixes.
Changes:
Fixed an issue where the player sprite would move up/down the screen
as you zoomed in/out.

Fixed animations being deletable (to delete an animation, delete the
parent sprites. We do this to make the management of those sprites
easier, we can rethink this if we ever need to).

Fixed a but where terrain was being initialized as flat ground
instead of empty.