ResourceImporter ID changes.
Changes:
Fixed a bug where IDs were being mishandled. On save, we were re-
numbering each object based on its order in the relevant map (we
were iterating maps and incrementing an int, starting at 1).
If a sprite was a part of an animation or set, this would
put some other sprite in its place.
From old comments, it seems like we did this to avoid IDs balooning
to very large values (if we always start from 1 and tightly pack
instead of marching forward like IDPool does, we end up with smaller
numbers). This only matters for increasing json size though, which
we've decided not to care about. Instead, we've gone back to using
the numericID that the object is assigned when its created, which
gives animations and sets a stable ID to reference (until the
object is deleted, at which point it should be cleared from all
uses).

Fixed a bug where the engine assumed that entries in the json would be
listed in the same order as their numeric ID. This assumption was
implicit in our use of emplace_back to add elements to the vectors,
instead of first parsing their numericID and adding them to the
associated slot.

Added a ReuseLowest strategy to IDPool. The engine still uses MarchForward,
but now the importer can reuse the lowest free IDs and keep the ID space
tight. This helps us in both the importer since we no longer have to pre-
allocate large vectors, and in the engine since the ID space won't grow
as we add/remove/add resources.

Added missing ID freeing to each model.