Fixed mouse picking crash when zoomed out.
Changes:
Added std::optional to screenToWorldRay-based transforms, for
when the mouse ray doesn't intersect the camera view bounds
(thus we can't reverse the ray). This can happen when zoomed
out.
An alternative would be to make zoom move the camera in/out and
expand the view bounds, but the view bounds have networking
implications so this seems better.