(Safety) Started adding migration tool.
Changes:
Split the persisted component lists in two separate database
columns, so we don't run into issues where a project AND the
engine both updated their lists at the same time, so it's
impossible to migrate the list as a whole. Instead, we can
run an engine-only migration path and a project-only migration
path.

Fixed how we were initializing the in-memory DB, apparently we
weren't loading anything from the database file into it prior.
Now it gets loaded at startup using the backup interface.

Added a -1 to the file DB's backup, so it actually copies all
pages.

Added a "versions" table, to track the version of each set of
data.

Added logic to check versions and refuse to start if they
don't match.

Switched some interfaces to use std::span. We weren't using it
before because it's a small performance regression on windows,
but we only really care about performance on the linux server.

Started building the migration tool.