Migration example works.
Changes:
Changed some db binds from casting to Uint32 to casting to int.
Likely we casted to Uint32 because that's the base type of the
variable, but the field in the db is an integer.

Changed the rest of the Database functions to use spans instead
of ptr + size.

Changed the Database::iterate functions to have all of their
columns indexed at the top. This gives us a clear index -> label
reference, hopefully avoiding bugs.

Migration example worked. Commented it out since it's expected to
break (whenever we actually need to do a migration, we can use it
as a starting point).