Item code quality improvements.
Changes:
Split ItemProperties into engine/project lists, like the other
type lists. It seemed to previously be following the enum pattern
of "put it all in the project", but that pattern only exists because
enum lists can't be joined. In any case where we can handle the engine's
part of a list in the engine code, we should.

Added a comment to ReplicatedComponent's size check. It seems like we
only limit the size to prevent data size from growing. That's probably
a good idea (every component taking an extra byte would be a decent
amount of data), but it should be clear that it isn't a hard limit.