Add item flag map and enum
Item bug fixes, improvements. Other bug fixes.

Changes:
Added ItemDataBase::loadItem(), moved createItem() and updateItem()
to Server::ItemData. "load" is the action that the client should always
be taking, loading exactly the data that the server sends it. The
client cache and server database also need to be loaded from. The
server needs to create and update items, though.

Removed Client::Simulation's item update signal. It was created before
we added create/update signals to ItemDataBase itself, and we just
forgot to switch things over.

Fixed an issue where ItemCache and Database weren't saving the item's
version. Likely we had the version in the Item definition before, and we
forgot to separately persist it when we moved it out. Also moved out
item init scripts to a separate map, to let items be packed more closely
together.

Removed the engine's item properties type list and moved the item
description into the Item class. All items have an Examine interaction so
having description be built in makes sense, and removing the engine type
list simplifies persistence and eventual migrations.

Renames variables in and around the database for more consistency. For
serialized data, we now call it "serializedXyz".

Moved lua environment library init into EngineLuaBindings. This lets us
do all the environment setup in one constructor, instead of splitting it up.

Switched some init script handling to use std::string_view instead of
constructing intermediate structs.

Fixed items being consumed on failure when trying to combine items if one
of the slots became empty or one of the items is deleted (item IDs are normally
never freed, but maybe the database was modified or something).

Moved the consts in Item.h to the middle of the class definition so it's easier
to see the owned data.