Removed AVSequences.
Changes:
Removed all references to AVSequences and AVEffects. This was a
failed design path, and we're going to go in a different direction.
The key takeaway is that there shouldn't be some separate effects
manager. Effects, like graphics, should be tied to the entities that
they're applied to. E.g. the sorter should be looking at an entity's
Buffs component to find buff effects, not a separate place.
If an effect needs to move around independent of any entity, a local
entity should be made to facilitate this.