Add component destruction replication, bug fixes.
Changes:
Updated entt, updated observers to new pattern.

Added replication for component destruction. Now if a
component is added to the Observed list, its destruction will
be replicated. If the whole entity is destroyed, individual updates
won't be sent, just the EntityDelete.

Refactored how we handle re-initializing NCEs. We previously
were trying to strip the components and clear the locators,
likely to keep the same entity ID. They were getting a new
ID on re-init anyway though (probably a mistake, we started
using createEntity) and it was fine, so we now just destroy
the entity and re-init on the next tick.
This fixes an issue where an entity would briefly exist with
no components (we assume all entities have Position).
This also caused problems with the new component destroy
observer, where the component destruction was being observed when
we only intended an EntityDelete to go out.

Fixed an item creation crash caused by our recent addition of the
null item to the maps.

Fixed an issue where we were failing to remove destroyed entities
from the collision locator, since they were already destroyed by
the time the callback tried to check for Collision. This was causing
phantom desynced collisions.