(Safety) Added cast cooldown handling.
Changes:
Added CastCooldown. It has a few important details:
It isn't created on spawn, we create it when the entity first
casts a castable. It never gets deleted after, so there isn't
much difference.
When it's constructed, we set it to the current time. Whenever
it's observed, we update it to the current time and remove any
expired cooldowns (e.g. we lazy-update it).
When it's loaded, we use SaveTimestamp to account for lost time.
We send initial state in a custom way, similar to Inventory.
This is because we can send the whole list, then the client can
keep it up-to-date based on CastStarted and CastFailed messages.

Added SaveTimestamp, for lazy-timer-based components to use to
catch up to the tick where the entity was saved.

Removed two component EnttObserver. The "store a bool when the
first component is constructed" approach means the entity will
never leave the storage if the second component is never
created. This is fine if you expect both to always eventually
be added, but we're better off just checking for both and
clearing the storage.