(Safety) Casting graphics are working.
Changes:
Entity graphic sets now contain 8 graphics for each graphic type
(1 for each Rotation::Direction). Previously we added a new type
for each direction, but this made it a pain to pass all of the
rotations into e.g. a Castable for a "casting" graphic. Now,
you just pass the type and use the entity's rotation to index
into a graphic.

Client::CastHelper now uses the player entity as the caster for
all casts instead of accepting an entity. The "castXYZ" functions
are only used by the client to start a cast, so they should only
ever have the client entity as the caster.

Refactored the animation logic to have a generic function that
should work for any graphic type, since rotation is now separated.
It defaults to Idle if the desired graphic isn't present, so it
won't work if you want some other default.

Moved the cast helper params back into the client and server
headers, since the client doesn't need 2 of the fields.

Turned DisplayStrings into a namespace, so we can add to it in-
place in other headers.