Added A/V entity support.
Changes:
Added A/V entities, for use as e.g. projectiles. Started
referring to normal entities as "server-synchronized" in
cases where we need to differentiate.

Fixed casts being killed client-side if the same cast is
re-attempted during casting.

Fixed failed casts not printing to the chat. Now there's a
signal in Simulation for UI to subscribe to for server
failures, and a helper for printing the failure types that
CastHelper returns.

Added GraphicHelpers so we can reuse some logic for A/V entities.

Animations now don't reset if you switch between directions of
the same type, so you can e.g. walk around smoothly.

Fixed an issue with animations starting on the wrong frame.

Fixed a ResourceImporter issue where generated sprite sheets
would calculate the sheet size properly, but lay the sprites
out in an extra column to the right, causing them to be cut
off.

We now send CastFailed properly when an entity moves during
casting.

Fixed an issue where cast cooldowns would be applied, even
if the cast was canceled. GCD still applies.