Cast fixes, comment improvements.
Changes:
Fixed an issue where item interactions were being blocked by
target range requirements when the interaction didn't require
a target.

Fixed an issue where we were taking const Vector3& for
targetPosition, but not initializing it. Now we take a value
that's initialized to a known out-of-bounds state. If a castable
requires a target position but you don't provide one, the shared
check will see that the position is out of bounds and fail.