A/V entity bug fixes.
Changes:
A/V entities are now considered to be "in front" of other
sprites, if their boxes are intersecting. This may look weird
sometimes (A/V entity is fully inside another object), but
generally should look better than constantly flipping sort order.

Fixed an issue where AVEntity::Behavior::CurrentPosition was missing
from some switch statements.

If we fail to find the target or valid graphic state of an A/V entity
during the update process, we now destroy the entity. This fixes
the issue of A/V entities staying alive forever when they hit an
error state.