(Safety) Started adding animation loop controls.
Changes:
Reworked how A/V entity timing works to make it more accurate.

Added capability for setting an A/V entity phase's duration to
-1 to make it last until the animation has played through once.
This can have visual issues if your animation is set to loop
(since it's controlled at the sim tick rate, but the renderer
usually ticks more often), but should be fine with animations
set to not loop.

Added ResourceImporter support for setting an animation's loop
behavior. Still need to add to engine.