Cast feel fixes.
Changes:
Client/server now block casts if the caster is currently moving.
This feels better than letting it get to the update, then killing
it.

Client cast failures are now signalled to the UI appropriately,
since the server doesn't send a message for the client's own entity.

Updated the cast messages to be more generic.