Improved cast feel after movement.
Changes:
Changed Client's CastHelper path to use a public "queue"
interface instead of immediately validating the cast.
This makes it so the tick's movement can happen before
the validation, so if you try to stop moving and cast on
the same tick, it'll work.

Castables now have an isTickSynchronized flag. If true,
the server will sort/queue the cast and process it on the
associated tick. This fixes the "cast failed because input
message wasn't yet processed" problem, while leaving a way
for things that should be instant (e.g. Examine) to still
be so.

Added a few headers to the PCH (found using visual studio's
"Build Insights" feature). This dropped build times by quite
a bit.