LoS checks for castables are working.
Changes:
Castable::TargetToolType -> TargetType.

Changed the "None" castable target tool type to more explicit
types. "None" didn't work in the case where a target entity was
provided, but the cast didn't care about it. We don't know whether
the cast callback is going to use it or not so we have to validate
it and fail on LoS/distance checks, but the castable may be completely
independent of any target, making it weird to fail.

Updated some raycast-related comments to mention "t values". This
seems to be accurate, and a more intuitive way to think about
what's going on.

Updated Server::CastHelper and CastSystem to check LoS. Clients don't
check LoS, mostly because other games don't seem to check it client-
side. Presumably, doing so causes failures when the spell would
actually succeed server-side. The cost of skipping a client-side
check is just a quick animation hiccup, which players are used to.

Added range/LoS checks right before a cast is finished, to make
sure it's still valid.

Fixed BoundingBox::intersects(start, end). Since an upper bound is
given, we need to pass 1 as tMaxBound and keep the direction
vector non-normalized (so that at t==1, we're at the end point).

Added more checks to castable creation, moved ID from a parameter to
just validating the value in the struct (it's nicer to see the value
at the top when you're defining the struct, instead of at the bottom
when you're adding it).

Added a raycast function to CollisionLocator. It's rough, we need to
refactor it and add the other versions still.

Added a const normal() function to Vector3. Maybe it's fine to have
both it and normalize(), or maybe we should eventually get rid of
normalize() and just use normal(), since it matches the other
functions better.

Changed ClientEntity/NonClientEntity in CollisionLocator to instead be
Static/Dynamic entity, which seems to be much more useful. We originally
threw the tags together very quickly, so it makes sense that we didn't
pick the most ideal pattern.