Server receive optimization.
Changes:
Refactored ClientHandler/Client so the client timeout is checked by
the disconnect thread instead of the receive thread. This lets us
avoid making OS calls to receive if the client doesn't have data
waiting.
This opens up the scenario where we could have checked the timeout
and avoided queueing data to be sent to an unresponsive client,
but that isn't a big deal since it will still get disconnected
on the next serviceClients() loop.