Fixed static entities not having collision.
Changes:
Fixed an issue where all entities were being added as dynamic.

The CollisionLocator::updateEntity interface was cleaner prior,
when we passed in the CollisionObjectType directly. The logic
required to check for input at all the callsites was a bit too
disconnected and ripe for future bugs though, so we made the
interface a bit messier to bake the logic into the function.

This does mean that, if we need to add more entity collision types
in the future, we'll either need to add more parameters to identify
the entity to the locator, or make it a CollisionObjectType param
again and have some other means for keeping the callsites in sync.