Reworked entity collision.
Changes:
Reworked entity collision. Entities now have a CollisionBitSets
component, with configurable collisionLayers and collisionMask.
This lets you modify an entity's collision behavior from a script,
and lets the project add collision layers.

Changed a couple functions with lots of parameters to use a struct.

Added entity exclusion list to raycasts. This lets us do a
raycastAny for LoS, and will be useful in other situations.

Entity collision now is properly commented. The intent before was
for it all to match the default graphic, but that was only in
comments. The code actually was matching it to the current
graphic set's idle south, which is what we now want.