Separated item string ID from display name.
Changes:
Item display names and string IDs are now independent (we used
to derive string ID from display name). It's useful at the project
level to be able to have a nice long display name, with a short,
concise string ID.
We may want to do a similar thing in the future to the resource
importer-related classes (sprite, etc). There's a bunch of ways to
do it (remove display name entirely, or get string ID from file name,
etc), not sure what would be best. But the current setup isn't
very useful.