Refactored ItemHandler, EnginePersistedComponent.
Changes:
ItemHandlers -> ItemHandler, now we only have a single script that
handles all items for an entity. This lets us key off the item's
properties, instead of needing to know exactly what items are
going to be used.

Moved EnginePersistedComponent out into a separate header with
ProjectPersistedComponent, since both are able to be handled
by the engine. This lets the project not have to see the
ugly library code that makes the type list work.

Moved SaveTimestamp to server, since the client doesn't use it.